+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 
+
+### Added
+
+- The `wait` subcommand. This allows you to wait for all tasks in the default queue/ a specific group to finish.
+    On top of this, you can also specify specific tasks ids.
+
 ## [0.8.2] - 2020-11-20
 
 ### Added